+++ castle/cms/static/plone-logged-in-compiled.js
@@ -100866,5 +100866,5 @@ require([
   }
 });
 
-define("/Users/brian.duncan/castle.cms/castle/cms/static/plone-logged-in.js", function(){});
+define("/Users/brian.duncan/castle-instances/pages-to-folders/castle/cms/static/plone-logged-in.js", function(){});