+++ CHANGES.md
@@ -4,7 +4,8 @@ Changelog
 3.0.0b129 (unreleased)
 ----------------------
 
-- Nothing changed yet.
+- Create Content Redirectable behavior so Dexterity objects can act as redirects
+- Create an action to convert pages into folders with the original page as the default view
 
 
 3.0.0b128 (2023-11-20)